refactor: move deselect filter test to filter-segment.e2e.ts
Moved the 'Deselecting a segment clears all active filters' test from
system-segments.e2e.ts to filter-segment.e2e.ts since the behavior
applies to all segments, not just system segments.
